# feat(perf): Remove additional function call to improve performance