<!---
  Include the number of the issue addressed by this PR above if applicable.
  PRs for code changes without an associated issue *will not be merged*.
  See CONTRIBUTING.md for more information.
-->
Resolves #
Will add document to reproduce benchmark

# Description

Query
```
mf query --metrics orders --group-by customer__customer_name --explain  | grep -i completed
mf query --metrics customers_with_orders --group-by customer__customer_name --explain  | grep -i completed
mf query --metrics orders --group-by customer__customer_name --explain  | grep -i completed
mf query --metrics food_revenue --group-by order_id__customer__last_ordered_at__year --explain  | grep -i completed
mf query --metrics revenue_growth_mom --group-by order_id__customer__first_ordered_at --explain  | grep -i completed
mf query --metrics revenue_growth_mom --group-by metric_time --explain  | grep -i completed
mf query --metrics customers_with_orders --group-by customer__customer_name --explain  | grep -i completed
```

## Before Optimization
╰─ ./bench.sh
✔ Success 🦄 - query completed after 0.20 seconds
✔ Success 🦄 - query completed after 0.21 seconds
✔ Success 🦄 - query completed after 0.20 seconds
✔ Success 🦄 - query completed after 0.47 seconds
✔ Success 🦄 - query completed after 0.28 seconds
✔ Success 🦄 - query completed after 0.20 seconds

## After Optimization
╰─ ./bench.sh
✔ Success 🦄 - query completed after 0.12 seconds
✔ Success 🦄 - query completed after 0.12 seconds
✔ Success 🦄 - query completed after 0.13 seconds
✔ Success 🦄 - query completed after 0.15 seconds
✔ Success 🦄 - query completed after 0.18 seconds
✔ Success 🦄 - query completed after 0.12 seconds

<!---
  Provide context for the Pull Request here, including more details on what
  is changing and why. Add any references and info to help reviewers
  understand your changes, such as any tradeoffs you considered, and the local
  test process you followed.
-->

<!--- 
  Before requesting review, please make sure you have:
  1. read [the contributing guide](https://github.com/dbt-labs/metricflow/blob/main/CONTRIBUTING.md),
  2. signed the [CLA](https://docs.getdbt.com/docs/contributor-license-agreements)
  3. run `changie new` to [create a changelog entry](https://github.com/dbt-labs/metricflow/blob/main/CONTRIBUTING.md#adding-a-changelog-entry)
-->